SimpleGANTrainer verified working!!
Moved testing into its own folder, and created simple program to test SimpleGANTrainer. It trains an extremely simple GAN for a specified number of epochs and outputs the discriminator's loss_by_epoch graph.

(The GAN performs very poorly, but this is because it is a poorly-designed GAN made mostly for the sole purpose of existing)