Added soft_save and soft_load functionality (and all associated subfunctions). All Trainer objects can now save and load all information stored in:  - stats  - in_functions  - loss_functions  - model state dicts  - optimizer state dicts  - totrain
With this functionality, Trainer objects can be saved to disk in their entirety, and can resume training when loaded from disk.

Note that this depends on pickle's functionality for saving and loading functions and classes. In particular, successfully loading custom functions and objects requires them to be defined in the current scope (so, successfully loading a totrain object requires importing the file where it's defined!). Furthermore, because soft_save only saves the model and optimizer state dicts, the new Trainer object must be instantiated with the correct model and optimizer objects.